Added advanced tagging and suggestion command
I modified the "For (Your name here)" in the embeds so now user feels a more personal experience.

Also, I added a new suggest command so people may suggest changes or other things to change the server Mirayoki is in and vote with a reaction like a poll.

Next up, create a economy system using MongoDB database wish me luck!