Spanish Expansion + Dev-Only Functions
This update adds more bilingual information in commands, adds an uptime clock in debug commands, creates a command to report bugs/errors and opens up a new feature, Dev-Only Powers, these are measured on Clearance Levels, (From I to X) that can only be changed from the master Database manually, this for Slash Studio Staff that may require different permissions to do different things.

Also a new log appears on the console with the current date and time when deploying Mirayoki to measure the time needed to be deployed.