Adding a license, updating packages, coinflip command and better logic.
This commit adds a License that will regulate how Mirayoki can be used by developers who would like to make a fork of it, a better logic for developer authentication and a brand new coinflip command, that will make you sweat with those bets with your friends.

Also made some changes to the package.json file and updated packages that Mirayoki uses to have the best performance, security and speed as always.