Improved Bot ID handling with environment variables
Often, when the user runs an administrative command that affects other users directly, the bot usually looks up the mentioned user to see if the command is not being used against the bot itself, as it may create a lot of errors and a possible crash.

Before, the bot used to check this using nothing but its ID as a string that was placed in an if statement.

Now, it uses environment variables to define the Bot's ID, which results in a more easy way to reuse the ID string, without having to copy/paste it every single time.