Migration to v13 Started
This is such a disaster, with the heavy task of having to rewrite Mirayoki all over again in order to support Slash commands and be up-to-date with the latest API, this is the first step of many to come. This will be no easy task. But the results will be amazing.