Starting database implementation
Mirayoki relies heavily on the MongoDB in order to display content that is unique to a user (Language, Economy and About me info)

As part of v13, the handler for MongoDB requests had to be rewritten in order to make it faster and more stable.

It also updates packages, improves console logs and the /ping command.