/clear, /about, RP, and 'Severe Error' help.
This commit is a big one, first, a /clear command has been rewritten. It provides the same function as the original one, clearing up to 100 messages at once.

Then, an /about command was created which reflects the same information as the original one.

Following, RP has been reintroduced and optimized for v13.

And finally, in case there is a coding error that may not let Mirayoki continue with its regular tasks, a 'Severe Error' embed will now appear to request the user to report this immediately.