Added Continuous Integration and CodeQL Analysis
To ensure the security of code as we write it, now GitHub's CodeQL will run a check to make sure that everything on Mirayoki runs securely, without any kind of weak point.

Also, GitHub actions will initialize the project to make sure it is working as supposed to, it will only make sure that everything on the index file (Mirayoki.js) is starting up.