Improved GitHub Actions
Now, logs contain more descriptive messages, and the graph displaying the jobs looks way more cooler now

Also, adds the dynamic step of sending the Tweet with the update because Tweets may only be sent whenever a "core" update was made, and that is why there were two different workflow files, that would run both at the same time and it would be very confusing.

In a few words, labels and descriptions for logs have been improved and the Tweet job has been improved to recognize when a "core" update is made.