Improved long quote handling + GitHub Actions Deployment
This commit adds a trim option for extraordinary long quotes. This was done to avoid Mirayoki to crash because of an invalid API request caused by a very long quote. Now, instead of failing, a very long quote will be shortened to 1022 characters and will have some points "..." if the quote exceeds this amount.

This commit also enables the deploying of Mirayoki from GitHub Actions and modifies some files regarding GitHub documentation.