Improving quoting authoring
This commit enables /quote to show the user who sent the quote in case they are not the assigned author. It also shows now the channel where the quote was sent from.

It also improves performance and updates packages, alongside with powering up the Process Manager by using the maximum instances possible to make sure there is no delay on the response of Mirayoki's servers.