Updated Heroku authentication
It's been a while, isn't it?

I have not been working on Mirayoki actively because I've been busy with finals and that stuff. Guess I'm free now and the school year is now really close to be over.

This commit updates the authorization workflow into the Heroku servers, after they suffered a breach they resetted everyone's passwords and OAuth authorizations. Hope it works now.

It also changes the email that can be used for contact me to my enterprise email.

And last but not least, it updataes the packages for maximum security and performance. And specifies the latest engines to be used by Heroku.