Lock and unlock channels
Users will now be able to lock channels to control the flow of messages in their servers. Doing so will restrict @everyone from sending messages and if the user decides to hide it, make it invisible to other members' eyes.

They may also unlock it, removing restrictions from @everyone regarding message sending. They may also choose to show the channel to all members if it was previously hidden.

This commit also updates engines and packages for optimized performance, and updates the CodeQL Action that is used to analyze Mirayoki's code.